Backend_API_Server/requirements.txt
Django>=4.2
python-jose>=3.3.0
python-dotenv>=1.0.0
# Add any other backend dependencies as needed (e.g., djangorestframework)
